Bump Postgres TLS cert lifetime
QE teams are using long living deployments (2+ years) and having short
lived certificates breaking their deployments. Re-issuing of those
certificates requires redeployment of some components to make use of the
new configmaps and secrets. In the common case the use of these secrets
is for short lived CI deployments where the lifetime of the certs does
not have effect.