Enable proactive scaler in staging
Since we use nodes with nvme in staging too, we need to enable proactive
scaler otherwise it take too much time to add a node. In addition, we
need staging to be configured the same way as prod so any test we run
there are more representative of production environment.

Signed-off-by: Hugo Ares <hares@redhat.com>