KFLUXINFRA-1131: Get namespaces and projects
Give the Konflux cluster roles the permission to get namespaces and
openshift projects. This will allow the user that has the role,
to see his/hers namespace/project on the OCP UI.
This permission would be also needed by "namespace-lister", which
lists all the namespace the user has get permission on.

Signed-off-by: Gal Ben Haim <gbenhaim@redhat.com>