namespace-lister: disable configmap name suffix hashing
Currently, when we generate the configmap for the proxy, kustomize will
automatically append a hash to its name.  This breaks the proxy
deployment, since it looks for a configmap with a specific name.

We can fix this either by disabling name suffix hashing or by injecting
the name via a replacement.  The former is easier to implement, so opt
to implement that method.

Signed-off-by: Andy Sadler <ansadler@redhat.com>