deploy namespace-lister to production
The deployments for namespace-lister appear to be stable on the relevant
staging cluster (`stone-stg-rh01`), so we can begin to roll this out to
the relevant production clusters.

For now, the only cluster we wish to deploy to is `kflux-prd-rh02`,
since the component `konflux-ui` has not been deployed to any other
production clusters yet.  All other clusters affected by the
`konflux-public-production` overlay will see no new resources added to
the cluster.

Adding namespace-lister to konflux-ui's production proxy config will
come in a separate PR.

Signed-off-by: Andy Sadler <ansadler@redhat.com>