Add per-platform instance tags to staging
When determining if a particular VM instance can be created, the VM
flavor's maximum number of instances is compared to the current number
of instances with the same instance-tag to ensure no more than the
maximum number of VMs are created. Since the current configurations
have all flavors using the same instance tag, the count of unique
instances (flavors, in this case) is wrong. Adding per-platform/flavor
instance tags should fix this.