Read Results API service metrics from api-for-watcher
Splitting the load of the Results API service, switched the
watcher/controller related request to a new pod which handles most of
the load. With the original service now used only by users, it gets much
less load (that was the purpose), but it triggers some false positives
from the monitoring service. Here we switch the ServiceMonitor to the
new pods running most of the load.