Run Mintmaker DependencyUpdateCheck cronjobs as non-root
Mintmaker SA has been granted the anyuid SCC becuase it needs to run its
subscription-manager pod as root[1]. A side effect is that now it tries
to run all of its pods as root by default. This is not secure and also
conflicts with the setting runAsNonRoot: true, resulting in failing
jobs. Explicitly run these cronjobs as a non-root user.

[1] https://github.com/konflux-ci/mintmaker/pull/222