PVO11Y-4717 Bump observability operator memory try #3
Increase the limit to 3Gi. This should allow observability-operator pod to survive the initial memory spike before the usage stabilises.

Sideffect of this is that this applies limits to ALL Operator pods, including prometheus-operator, perses-operator and webhook-admission.
This should not impact scheduling as limits are not considered during scheduling.

Existing proposal for configuring operator resource usage: https://github.com/operator-framework/operator-controller/issues/1507