Configure prod-p02 cluster to test static PVMs
Multi-architectural builds in the prod-p02 cluster are timing out due to
waiting for PPC VMs to be provisioned. To decrease the time needed to
complete the build, a set of static VMs is provisioned so that the only
wait time is for if a VM is at its current concurrency limit. These
static VMs are currently prefaced with "test" so testing them will not
interfere with current builds.