etcd-shield: deploy to staging
Overview
========

etcd-shield is a load-shedding tool designed to provide a more
consistent experience to users when an outage is occurring.  When a
configured alert is firing, it denies the admission of pipelineruns into
the cluster.

Currently, it's configured to watch for etcd's on-disk capacity (hence
the name).  The database capacity limits are not stored as prometheus
time-series entries, so we're forced to hard-code the current 8GiB limit
in the PrometheusRule resource.

In the future, we could add more alerts to match on and more resources
types for this service to deny.

Alert configuration
===================

The alert is configured to trigger at 95% capacity, and hold that
trigger until capacity drops below 80%.  This is done to provide a
buffer between when we stop allowing new resources and when we start
providing them again.  We'd like to ensure that when we allow resource
admission again, it isn't going to immediately begin denying them;
having a gap of 15% will buy the cluster some breathing room.

Of course, these capacity percentages can be changed if we need to.

Signed-off-by: Andy Sadler <ansadler@redhat.com>