Increase defrag threshold to 30%
Having a threshold of 15% is too aggressive causing unnecessary
defragmentation when DB size is < 4GB. At 4 GB, 15% means defrag will
happen as soon as 607MB is reclaimable. There is not point defragmenting
for 607MB when DB size is at 50% capacity.

Increase the threshold to 30% to reduce defrag when DB size in not even
50%. When DB is larger 60-70%, this is where we need to more
aggressively defrag and this is happening due to other condition where
we do it if 1GB is reclaimable.

KFLUXINFRA-1644

Signed-off-by: Hugo Ares <hares@redhat.com>