KONFLUX-1624 - Deploy insights' ClusterSecretStore to only one cluster
The insights-secret-store was being deployed to all the production
clusters but it was failing in the recently created ones (prd-rh03 and
rhel-01) as the required secret to access Vault was not found.

After closer examination, it was found that only one production system,
prd-rh01, is using this cluster secret store as the other ones do not
even contain the target namespaces.

This commit deploys the insights-secret-store only to prd-rh01.