KONFLUX-1624 - Move base into production folder
When creating the CLusterSecretStores for the 'RH' instances, there is
no customization for any other environment than production. As such,
move the 'base' folder into the 'production' one as a first step. An
upcoming commit will take care of customizing the production instances
so that only CLusterSecretStores required in each instance are created.