etcd-shield: fix metrics in staging
Metrics for etcd-shield were misconfigured in two ways.  First, the
labels selectors were incorrect and not applied to the relevant service.
Applying the correct labels and label selectors allows the
servicemonitor to pick up the correct metrics endpoints.  Second,
because etcd-shield is run not as a user workload but as a core service,
the prometheus-k8s service account in the namespace
`openshift-monitoring` needs to be given appropriate permissions to look
at resources in the `etcd-shield` namespace.  This is fixed by adding a
role and a rolebinding.

Signed-off-by: Andy Sadler <ansadler@redhat.com>