o11y: Split MonitoringStack by environment
This is in attempt to speed up the pipeline whenever changes to this
MonitoringStack definition is made. This is especially crucial for if we
need to revert a breaking change to this definition.
Additional, this allows us to test changes to this definition in dev or
stage first before applying them to production.