backup: remove host cluster configs from prod
Remove the backup configurations referencing toolchain-host-operator
from staging. This operator has not existed for a while, and the
manifests here no longer serve any purpose.

Furthermore, the host clusters are being decommissioned, and we're
removing the concept of a host-cluster entirely from this repository and
from our deployments of konflux.

Part-of: KFLUXINFRA-2240
Signed-off-by: Andy Sadler <ansadler@redhat.com>