remove host-specific components
With the decommissioning of the host clusters, we no longer need to have
components specific to the host clusters.  Remove the applicationsets
only deployed to the host clusters, and remove their references from the
overlays.

We cannot merge argo-cd-apps/base/member and all-clusters, since we
still have the EaaS clusters.  These clusters have components that
should not be deployed to the member clusters, and vice-versa.

Part-of: KFLUXINFRA-2240
Signed-off-by: Andy Sadler <ansadler@redhat.com>