ci: Add GH workflow to ensure pipelines new index image is an upgrade
Pipelines nightly builds use an incrementing build number to determine
when to upgrade. If an index is used which has a lower build number than
the currently-deployed build, the operator silently skips the upgrade.

This adds a CI job which runs any time a pipelines index is changed and
warns the the PR author when a new index will not result in an upgrade.