KFLUXINFRA-2511 - Remove OLM ESO from stone-stage-p01
Remove the OLM External Secrets Operator (ESO) from stone-stage-p01
by creating an empty kustomization overlay to prevent deployment to
this cluster while maintaining normal deployment to other staging
clusters. The goal is to replace the OLM ESO with the Helms Charts
version, change that will be done in a follow up commit.