KFLUXINFRA-2511 - Remove OLM ESO from stone-stage-p01
Add a merge generator to ApplicationSet with cluster-specific override
for stone-stage-p01 and create an empty kustomization overlay to prevent
deployment to this cluster while maintaining normal deployment to other
staging clusters. The end goal is to remove the OLM External Secret
Operator (ESO) from this cluster to replace it with the Helms Chart
version, change that will be done in a follow up commit.