KFLUXINFRA-2511 - Add Helm Charts based ESO deployment
Add the Helm Charts based deployment of External Secrets Operator to
one staging cluster, 'stone-stg-p01', as a first step to validate the
migration does not produce any issue. The version deployed is the same
as the OLM version deployed in other clusters (0.11.0) to minimize the
possibility of introducing any regression.