KFLUXINFRA-2511 - Add Helm Charts based ESO deployment
Add the Helm Charts based deployment of External Secrets Operator to
one staging cluster, 'stone-stg-p01', as a first step to validate the
migration does not produce any issue. The version deployed is the same
as the OLM version deployed in other clusters (0.11.0) to minimize the
possibility of introducing any regression.

Pulling the ESO Helm Chart from a remote repository fails with this error:

  "Error when running kustomize build for components/external-secrets-operator/helm-charts:
   Error: failed to untar: a file or directory with the name
   /home/runner/work/infra-deployments/infra-deployments/components/external-secrets-operator/helm-charts/charts/external-secrets-0.11.0/external-secrets-0.11.0.tgz
   already exists""

After several unsuccesful workarounds and given the recommended way of
handling the charts is by 'vendoring' them [1], the ESO Helm Charts are
added to this commit.

[1] https://github.com/kubernetes-sigs/kustomize/blob/master/examples/chart.md#best-practice