kueue: add mintmaker resource to staging
Recently, we've been seeing etcd get alarmingly full on stone-prd-rh01.
Sometimes it's enough for etcd-shield to kick in, which causes a build
outage that impacts our users (and may impact our SLOs).  We need to
start looking into methods of reducing the number of resources stored in
etcd.

One of the causes behind our occupancy issues is mintmaker's update
checks that run every four hours.  Between the pipelineruns that
mintmaker generates directly and the on-pull-request pipelineruns
triggered on downstream repositories, we're generating enough load for
etcd to fill up.  In particular, we've seen at least 1.5 MILLION events
on the cluster just from these workloads.

To address this issue, we need to reduce the amount of occupancy in
etcd.  We're currently pursuing multiple methods, but one way to prevent
mintmaker from overloading our clusters is to get mintmaker's
pipelineruns to participate more in kueue's resource sharing.  Kueue
already imposes a limit on the number of pipelineruns the clusters can
run at any point in time, but we can teach it to have a limit on the
number of concurrent mintmaker pipelineruns that can run at any time.

Add a new resource type to kueue, "mintmaker", that's set on pipelines
run in mintmaker's namespace. With this, we can put an upper bound on
the number of mintmaker pipelines allowed to run simultaneously; for
now, set an upper bound of 200.

Part-of: KFLUXINFRA-2526
Signed-off-by: Andy Sadler <ansadler@redhat.com>