kueue: skip old platform param if new param exists in production
If both the old and the new platform params exist in the pipelinerun, we
should only pull from the new platform params. The old parameters are
likely going to be something we can't interpret with CEL expressions if
the new paramter is set. To avoid issues, only read from the old
expressions if the new expressions are not set.

See KONFLUX-9917 for issues that have already cropped up in our
production clusters.

Fixes: KONFLUX-9917
Signed-off-by: Andy Sadler <ansadler@redhat.com>