Push ETCD defrag threshold to 75%
The goal of this change is to reduce the defrag we do to the minimal, we
want to avoid at all cost unneeded defrag causing API server request to
fail. After monitoring for few hours, I see that we still trigger
unneeded defrag so push the limit to 75%, which is 5% below etcd shield
threshold.

Signed-off-by: Hugo Ares <hares@redhat.com>